pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quiver"
version = "0.1.0"
description = "Intelligence wargame management system"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3,<3",
    "flask>=3.0,<4",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/quiver"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
